+++ src/Message/ServerRequest/Stringable.php
@@ -3,19 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Message\ServerRequest;
 
-use Innmind\Http\{
-    ServerRequest,
-    Message\Method,
-    Message\Environment,
-    Message\Cookies,
-    Message\Query,
-    Message\Form,
-    Message\Files,
-    ProtocolVersion,
-    Headers,
-    Header,
-};
-use Innmind\Url\Url;
+use Innmind\Http\ServerRequest;
 use Innmind\Filesystem\File\Content;
 use Innmind\Immutable\{
     Sequence,
@@ -27,80 +15,19 @@ use Innmind\Immutable\{
  */
 final class Stringable
 {
-    private ServerRequest $request;
-
-    public function __construct(ServerRequest $request)
+    private function __construct()
     {
-        $this->request = $request;
     }
 
-    /**
-     * @psalm-pure
-     */
-    public static function of(ServerRequest $request): self
-    {
-        return new self($request);
-    }
-
-    public function url(): Url
-    {
-        return $this->request->url();
-    }
-
-    public function method(): Method
-    {
-        return $this->request->method();
-    }
-
-    public function protocolVersion(): ProtocolVersion
-    {
-        return $this->request->protocolVersion();
-    }
-
-    public function headers(): Headers
-    {
-        return $this->request->headers();
-    }
-
-    public function body(): Content
-    {
-        return $this->request->body();
-    }
-
-    public function environment(): Environment
-    {
-        return $this->request->environment();
-    }
-
-    public function cookies(): Cookies
-    {
-        return $this->request->cookies();
-    }
-
-    public function query(): Query
-    {
-        return $this->request->query();
-    }
-
-    public function form(): Form
-    {
-        return $this->request->form();
-    }
-
-    public function files(): Files
-    {
-        return $this->request->files();
-    }
-
-    public function asContent(): Content
+    public function __invoke(ServerRequest $request): Content
     {
         $status = Str::of("%s %s%s HTTP/%s\n")->sprintf(
-            $this->method()->toString(),
-            $this->url()->path()->toString(),
-            $this->queryString(),
-            $this->protocolVersion()->toString(),
+            $request->method()->toString(),
+            $request->url()->path()->toString(),
+            $this->queryString($request),
+            $request->protocolVersion()->toString(),
         );
-        $headers = $this
+        $headers = $request
             ->headers()
             ->all()
             ->map(static fn($header) => $header->toString())
@@ -111,36 +38,39 @@ final class Stringable
             Sequence::lazyStartingWith($status)
                 ->append($headers)
                 ->add(Str::of("\n"))
-                ->append($this->bodyChunks()),
+                ->append($this->bodyChunks($request)),
         );
     }
 
-    public function toString(): string
+    /**
+     * @psalm-pure
+     */
+    public static function new(): self
     {
-        return $this->asContent()->toString();
+        return new self;
     }
 
-    private function queryString(): string
+    private function queryString(ServerRequest $request): string
     {
-        if (\count($this->query()) === 0) {
+        if (\count($request->query()) === 0) {
             return '';
         }
 
-        return '?'.\rawurldecode(\http_build_query($this->query()->data()));
+        return '?'.\rawurldecode(\http_build_query($request->query()->data()));
     }
 
     /**
      * @return Sequence<Str>
      */
-    private function bodyChunks(): Sequence
+    private function bodyChunks(ServerRequest $request): Sequence
     {
-        if (\count($this->form()) !== 0) {
-            return Sequence::of($this->form()->data())
+        if (\count($request->form()) !== 0) {
+            return Sequence::of($request->form()->data())
                 ->map(\http_build_query(...))
                 ->map(\rawurldecode(...))
                 ->map(Str::of(...));
         }
 
-        return $this->body()->chunks();
+        return $request->body()->chunks();
     }
 }